Crate snowcap_runtime

Runtime System

This system generates a virtual network inside GNS3, configures all nodes and performs the migration scenario, while monitoring the forwarding state. For simplified usage, check the function perform_migration.

Modules

checker
This module checks the paths if the conditions are ok

config
This module is responsible for parsing the config correctly

frr_conn
Utilities for telnet interactions

pcap_reader
Reads pcap files and extracts all udp packets with exactly 8 bytes, to interpret them.

physical_network
Physical representation of the abstract network from snowcap

python_conn
Utilities for telnet interactions to a Python Docker Container

Functions

perform_migration
Perform the migraiton