Struct snowcap_runtime::frr_conn::FrrConnection

pub struct FrrConnection { /* fields omitted */ }

Connection to FRRouting

This struct can be used to communicate with an FRR instance, running inside GNS3, using telnet. It can be used to configure and reconfigure the router, as well as to check its forwarding table and the result of simple traceroute commands.

All commands are synchronous and blocking. Non-blocking implementation could be used theoretically, but it was not worth the hassle. However, the blocking wait functions are all implemented using a busy loop with a very large sleep timer in between. This still allows many parallel connections to be established at the same time.

This struct does not implement Copy, Sync or Send, since it involves communicating with a stream from the OS.

Implementations

impl FrrConnection

pub fn new(port: u16) -> Result<Self, Box<dyn Error>>

create a new connection to a device

pub fn reconfigure(
&mut self,
expr: Vec<impl AsRef<str>>
) -> Result<(), Box<dyn Error>>

Reconfigure a specific option in the router. if the configuration needs to happen inside a nested group, use the first element in the expr to navigate into this position, and the last to set the actual configuration.

pub fn traceroute(
&mut self,
target: &IpAddr
) -> Result<Option<Vec<[u8; 4]>>, Box<dyn Error>>

Perform a traceroute, and get the result back, as a vector of IP addresses.

pub fn get_routing_table(&mut self) -> Result<RoutingTable, Box<dyn Error>>

Returns the current routing table

pub fn initialize_config(
&mut self,
router: &PhysicalRouter
) -> Result<(), Box<dyn Error>>

Apply an entire config to the router