Function snowcap_runtime::pcap_reader::extract_pcap_flows

pub fn extract_pcap_flows(
    filename: impl AsRef<str>
) -> Result<HashMap<u32, Vec<u32>>, Box<dyn Error>>

Read a pcap file, extract all packets, and return which flows and which sequence numbers of these flows have been seen