Function snowcap_runtime::pcap_reader::path_inference

pub fn path_inference(
    captures: Vec<(RouterId, RouterId, HashMap<u32, Vec<u32>>)>, 
    flows: &HashMap<(RouterId, Prefix), u32>
) -> HashMap<(RouterId, Prefix), HashMap<Option<Vec<RouterId>>, usize>>

Infer the path of all packets using the results from the pcap files

Input: Every element of the vector contains the following fields:

Link node A,

Link node B,

Table, mapping the flow id to the received sequence numbers (sorted) at this node.

The output is shaped as follows: For each flow id, we have a table, which maps a path (wrapped in an option to also incoperate when packets were dropped) to the number of packets that took this path.

TODO

This algorithm can be improved, if we would actively go through every capture in sequential order, removing sequence numbers that we have already seen.